fix: refactor hardhat config file
