+++ .env.example
@@ -28,7 +28,7 @@ MAINNET_RPC_URL=http://localhost:8545
 
 # RPC URL for Hardhat Network forking, required for running tests on mainnet fork with tracing (Infura, Alchemy, etc.)
 # https://hardhat.org/hardhat-network/docs/guides/forking-other-networks#forking-other-networks
-HARDHAT_FORKING_URL=
+HARDHAT_FORKING_URL=https://eth.drpc.org
 
 # https://docs.lido.fi/deployed-contracts
 MAINNET_LOCATOR_ADDRESS=0xC1d0b3DE6792Bf6b4b37EccdcC24e45978Cfd2Eb
