+++ tasks/index.ts
@@ -1,2 +1,3 @@
 export * from "./verify-contracts";
 export * from "./extract-abis";
+export * from "./solidity-get-source";
